Shows a list of roles on update and delete operation in Role Registry.

+++ app/src/main/java/com/ecc/hibernate_xml/app/RoleUiHandler.java
@@ -13,6 +13,8 @@ public class RoleUiHandler {
 	private static final String NAME_PROMPT = "Please enter the new role: "; 
 
 	private static final String NO_ROLES_MESSAGE = "There are no roles.";
+	private static final String NO_ROLES_TO_UPDATE = "There are no roles to update.";
+	private static final String NO_ROLES_TO_DELETE = "There are no roles to delete.";
 	private static final String CREATE_SUCCESS_MESSAGE = "Successfully created the role ID \"%d\" with \"%s\"!";
 	private static final String UPDATE_SUCCESS_MESSAGE = "Successfully updated the role ID \"%d\" with \"%s\"!";
 	private static final String DELETE_SUCCESS_MESSAGE = "Successfully deleted the role ID \"%d\"!";
@@ -42,30 +44,53 @@ public class RoleUiHandler {
 	}
 
 	public void update() throws Exception {			
-		Integer roleId = InputHandler.getNextLine(UPDATE_PROMPT, Integer::valueOf);
-		RoleDTO role = roleService.get(roleId);
+		List<RoleDTO> roles = roleService.list();
 
-		role.setName(InputHandler.getNextLineREPL(NAME_PROMPT, roleService::validateName));
-		roleService.update(role);
-
-		String successMessage = String.format(UPDATE_SUCCESS_MESSAGE, role.getId(), role.getName());
-
-		if (role.getPersons().size() > 0) {
-			String personIds = role.getPersons()
-				.stream()
-				.map(id -> id.toString())
-				.collect(Collectors.joining(", "));
-
-			successMessage += " " + String.format(AFFECTED_PERSONS_MESSAGE, personIds);
+		if (roles.isEmpty()) {
+			System.out.println(NO_ROLES_TO_UPDATE);
 		}
-		System.out.println(successMessage);
+		else {		
+			roles.stream().map(role -> role.toString())
+				.forEach(System.out::println);
+			System.out.println("-------------------");		
+
+			Integer roleId = InputHandler.getNextLine(UPDATE_PROMPT, Integer::valueOf);
+			RoleDTO role = roleService.get(roleId);
+
+			role.setName(InputHandler.getNextLineREPL(NAME_PROMPT, roleService::validateName));
+			roleService.update(role);
+
+			String successMessage = String.format(UPDATE_SUCCESS_MESSAGE, role.getId(), role.getName());
+			if (role.getPersons().size() > 0) {
+				String personIds = role.getPersons()
+					.stream()
+					.map(id -> id.toString())
+					.collect(Collectors.joining(", "));
+
+				successMessage += " " + String.format(AFFECTED_PERSONS_MESSAGE, personIds);
+			}
+			System.out.println(successMessage);
+		}		
+		System.out.println("-------------------");
 	}
 
-	public void delete() throws Exception {			
-		Integer roleId = InputHandler.getNextLine(DELETE_PROMPT, Integer::valueOf);
-		roleService.delete(roleId);
+	public void delete() throws Exception {	
+		List<RoleDTO> roles = roleService.list();
 
-		String successMessage = String.format(DELETE_SUCCESS_MESSAGE, roleId);
-		System.out.println(successMessage);
+		if (roles.isEmpty()) {
+			System.out.println(NO_ROLES_TO_DELETE);
+		}
+		else {		
+			roles.stream().map(role -> role.toString())
+				.forEach(System.out::println);
+			System.out.println("-------------------");		
+	
+			Integer roleId = InputHandler.getNextLine(DELETE_PROMPT, Integer::valueOf);
+			roleService.delete(roleId);
+
+			String successMessage = String.format(DELETE_SUCCESS_MESSAGE, roleId);
+			System.out.println(successMessage);
+		}		
+		System.out.println("-------------------");		
 	}
 }
